+++ components/command/command-controller/src/test/java/uk/gov/justice/services/components/command/controller/interceptors/CommandControllerInterceptorChainProviderTest.java
@@ -7,12 +7,14 @@ import static uk.gov.justice.services.core.annotation.Component.COMMAND_CONTROLL
 
 import uk.gov.justice.services.core.accesscontrol.LocalAccessControlInterceptor;
 import uk.gov.justice.services.core.audit.LocalAuditInterceptor;
-import uk.gov.justice.services.core.interceptor.InterceptorChainEntry;
+import uk.gov.justice.services.core.interceptor.Interceptor;
 import uk.gov.justice.services.metrics.interceptor.IndividualActionMetricsInterceptor;
 import uk.gov.justice.services.metrics.interceptor.TotalActionMetricsInterceptor;
 
 import java.util.List;
 
+import org.apache.commons.lang3.tuple.ImmutablePair;
+import org.apache.commons.lang3.tuple.Pair;
 import org.junit.Test;
 
 public class CommandControllerInterceptorChainProviderTest {
@@ -25,12 +27,12 @@ public class CommandControllerInterceptorChainProviderTest {
     @Test
     @SuppressWarnings("unchecked")
     public void shouldProvideDefaultInterceptorChainTypes() throws Exception {
-        final List<InterceptorChainEntry> interceptorChainTypes = new CommandControllerInterceptorChainProvider().interceptorChainTypes();
+        final List<Pair<Integer, Class<? extends Interceptor>>> interceptorChainTypes = new CommandControllerInterceptorChainProvider().interceptorChainTypes();
 
         assertThat(interceptorChainTypes, containsInAnyOrder(
-                new InterceptorChainEntry(1, TotalActionMetricsInterceptor.class),
-                new InterceptorChainEntry(2, IndividualActionMetricsInterceptor.class),
-                new InterceptorChainEntry(3000, LocalAuditInterceptor.class),
-                new InterceptorChainEntry(4000, LocalAccessControlInterceptor.class)));
+                new ImmutablePair<>(1, TotalActionMetricsInterceptor.class),
+                new ImmutablePair<>(2, IndividualActionMetricsInterceptor.class),
+                new ImmutablePair<>(3000, LocalAuditInterceptor.class),
+                new ImmutablePair<>(4000, LocalAccessControlInterceptor.class)));
     }
 }