+++ components/command/command-handler/src/main/java/uk/gov/justice/services/components/command/handler/interceptors/CommandHandlerInterceptorChainProvider.java
@@ -5,13 +5,14 @@ import static uk.gov.justice.services.core.annotation.Component.COMMAND_HANDLER;
 
 import uk.gov.justice.services.components.common.BaseInterceptorChainProvider;
 import uk.gov.justice.services.core.audit.LocalAuditInterceptor;
-import uk.gov.justice.services.core.interceptor.InterceptorChainEntry;
+
+import org.apache.commons.lang3.tuple.ImmutablePair;
 
 public class CommandHandlerInterceptorChainProvider extends BaseInterceptorChainProvider {
 
     public CommandHandlerInterceptorChainProvider() {
-        interceptorChainTypes().add(new InterceptorChainEntry(MAX_VALUE, RetryInterceptor.class));
-        interceptorChainTypes().add(new InterceptorChainEntry(3000, LocalAuditInterceptor.class));
+        interceptorChainTypes().add(new ImmutablePair<>(MAX_VALUE, RetryInterceptor.class));
+        interceptorChainTypes().add(new ImmutablePair<>(3000, LocalAuditInterceptor.class));
     }
 
     @Override