+++ components/event/event-processor/src/test/java/uk/gov/justice/services/components/event/processor/interceptors/EventProcessorInterceptorChainProviderTest.java
@@ -5,12 +5,16 @@ import static org.hamcrest.Matchers.is;
 import static org.junit.Assert.assertThat;
 import static uk.gov.justice.services.core.annotation.Component.EVENT_PROCESSOR;
 
-import uk.gov.justice.services.core.interceptor.InterceptorChainEntry;
+import uk.gov.justice.services.core.accesscontrol.LocalAccessControlInterceptor;
+import uk.gov.justice.services.core.audit.LocalAuditInterceptor;
+import uk.gov.justice.services.core.interceptor.Interceptor;
 import uk.gov.justice.services.metrics.interceptor.IndividualActionMetricsInterceptor;
 import uk.gov.justice.services.metrics.interceptor.TotalActionMetricsInterceptor;
 
 import java.util.List;
 
+import org.apache.commons.lang3.tuple.ImmutablePair;
+import org.apache.commons.lang3.tuple.Pair;
 import org.junit.Test;
 
 public class EventProcessorInterceptorChainProviderTest {
@@ -23,10 +27,10 @@ public class EventProcessorInterceptorChainProviderTest {
     @Test
     @SuppressWarnings("unchecked")
     public void shouldProvideDefaultInterceptorChainTypes() throws Exception {
-        final List<InterceptorChainEntry> interceptorChainTypes = new EventProcessorInterceptorChainProvider().interceptorChainTypes();
+        final List<Pair<Integer, Class<? extends Interceptor>>> interceptorChainTypes = new EventProcessorInterceptorChainProvider().interceptorChainTypes();
 
         assertThat(interceptorChainTypes, containsInAnyOrder(
-                new InterceptorChainEntry(1, TotalActionMetricsInterceptor.class),
-                new InterceptorChainEntry(2, IndividualActionMetricsInterceptor.class)));
+                new ImmutablePair<>(1, TotalActionMetricsInterceptor.class),
+                new ImmutablePair<>(2, IndividualActionMetricsInterceptor.class)));
     }
 }